pyproject.toml
[tool.poetry]
name = "lezyne-livetrack"
version = "0.1.0"
description = ""
authors = ["Alex Khaerov <hayorov@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
imap-tools = "^0.35.0"
redis = "^3.5.3"
python-dotenv = "^0.15.0"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py38']